+++ core/dbt/rpc/__init__.py
@@ -34,10 +34,3 @@ The general idea from a thread/process management perspective (ignoring the
   - `kills` all processes (triggering the end of all processes, right!?)
   - exits (all remaining threads should die here!)
 """
-from dbt.rpc.error import (  # noqa
-    dbt_error, server_error, invalid_params, RPCException
-)
-from dbt.rpc.logger import RemoteCallableResult  # noqa
-from dbt.rpc.task import RemoteCallable  # noqa
-from dbt.rpc.task_manager import TaskManager  # noqa
-from dbt.rpc.response_manager import ResponseManager  # noqa
